move edition derivation checks to common handler
